calimero_server::admin::utils
Module auth

Functions

construct_node_challenge

Constructs a node challenge message from the provided signature message.

decode_signature

Decodes a base64-encoded signature.

eth_message

is_older_than_15_minutes

Checks if a given timestamp is older than 15 minutes from the current time.

serialize_node_challenge

Serializes the node challenge into a JSON string.

validate_challenge

Validates the challenge by verifying the node signature and checking if it has expired.

validate_challenge_content

Validates the content of the challenge by checking the node’s signature.

validate_root_key_exists

Validates if the root key exists for the given request.

verify_eth_signature

verify_near_public_key

verify_node_signature

Verifies a node signature based on the type of wallet (NEAR, ETH, STARKNET).

verify_signature

Verifies the signature of a message using the node’s keypair.